Add Support for ddev-nextcloud-fpm

+++ src/Command/DeployGenerate.php
@@ -17,10 +17,19 @@ class DeployGenerate
     ): int
     {
         $ddevDockerCompose = Yaml::parseFile(getenv('DDEV_APPROOT') . '/.ddev/.ddev-docker-compose-full.yaml');
+
+        $projectType = $this->getProjectType($ddevDockerCompose);
+
+        if (!$projectType) {
+            $output->writeln('Project type not supported');
+        }
+
+        $projectTypeConfig = require __DIR__  . '/../ProjectType/' . $projectType . '.php';
+
         $ddevFilesDirs = $ddevDockerCompose['services']['web']['environment']['DDEV_FILES_DIRS'];
         // Get the upload dirs as an array of relative paths like:
         // ['web/sites/default/files'].
-        $ddevUploadDirs = array_map(function ($uploadDir) {
+        $ddevUploadDirs = empty($ddevFilesDirs) ? [] : array_map(function ($uploadDir) {
             return str_replace(getenv('DDEV_APPROOT') . '/', '', $uploadDir);
         }, explode(',', $ddevFilesDirs));
 
@@ -49,6 +58,7 @@ class DeployGenerate
                 'ddev_cron_job_day' => "{{ lookup('env','DDEV_CRON_JOB_DAY') | default('*', true) }}",
                 'ddev_cron_job_month' => "{{ lookup('env','DDEV_CRON_JOB_MONTH') | default('*', true) }}",
                 'ddev_cron_job_weekday' => "{{ lookup('env','DDEV_CRON_JOB_WEEKDAY') | default('*', true) }}",
+                'ddev_services' => array_keys($ddevDockerCompose['services']),
             ],
             'roles' => [[
                 'role' => 'ansistrano.deploy',
@@ -90,7 +100,7 @@ class DeployGenerate
             'name' => 'Log the output of stopping the current release',
             'copy' => [
                 'content' => "STDOUT\n{{ stop_result.stdout }}\nSTDERR\n{{ stop_result.stderr }}",
-                'dest' => '{{ ansistrano_release_path.stdout }}/ansible-ddev-stop-previous-current.log.txt',
+                'dest' => '{{ ansistrano_release_path }}/ansible-ddev-stop-previous-current.log.txt',
             ],
             'when' => 'current_release.stat.exists',
         ];
@@ -100,57 +110,67 @@ class DeployGenerate
                 // Trailing slash is important to copy the contents of src and
                 // avoiding a structure like `/files/files/files`, etc.
                 'src' => "{{ ansistrano_deploy_to }}/current/{{ item }}/",
-                'dest' => '{{ ansistrano_release_path.stdout }}/{{ item }}',
+                'dest' => '{{ ansistrano_release_path }}/{{ item }}',
                 'remote_src' => true,
             ],
             'loop' => '{{ ddev_upload_dirs }}',
             'when' => 'current_release.stat.exists',
         ];
 
-        // TODO: Inspect volumes to avoid hardcoding mariadb.
-        $afterUpdateCodeTasks[] = [
-            'name' => 'Copy volumes from the current release to the next one',
-            'shell' => 'docker container run --rm ' .
-            '--volume {{ ddev_project_name }}-{{ ddev_ansible_environment }}-{{ ansistrano_current_release_version.content | b64decode | lower }}-mariadb:/from ' .
-            '--volume {{ ddev_project_name }}-{{ ddev_ansible_environment }}-{{ ansistrano_release_version | lower }}-mariadb:/to ' .
-            'alpine:3.23.3 sh -c "cd /from ; cp -a . /to"',
-            'when' => 'current_release.stat.exists',
-        ];
+        foreach ($ddevDockerCompose['volumes'] as $volumeName => $volume) {
+            // Consider only volumes that start with the project name ignoring
+            // volumes like ddev-global-cache
+            if (!str_starts_with($volume['name'], getenv('DDEV_PROJECT'))) {
+                continue;
+            }
+            $name = str_replace(getenv('DDEV_PROJECT') . '-', '', $volume['name']);
+
+            $afterUpdateCodeTasks[] = [
+                'name' => 'Copy volume "' . $volumeName . '"(' . $name . ')  from the current release to the next one',
+                'shell' => 'docker container run --rm ' .
+                '--volume {{ ddev_project_name }}-{{ ddev_ansible_environment }}-{{ ansistrano_current_release_version.content | b64decode | lower }}-' . $name . ':/from ' .
+                '--volume {{ ddev_project_name }}-{{ ddev_ansible_environment }}-{{ ansistrano_release_version | lower }}-' . $name . ':/to ' .
+                'alpine:3.23.3 sh -c "cd /from ; cp -a . /to"',
+                'when' => 'current_release.stat.exists',
+            ];
+        }
         $afterUpdateCodeTasks[] = [
             'name' => 'Config the app to include environment and version',
             'shell' => 'ddev config --project-name "{{ ddev_project_name }}-{{ ddev_ansible_environment }}-{{ ansistrano_release_version | lower }}"',
             'args' => [
-                'chdir' => '{{ ansistrano_release_path.stdout }}',
+                'chdir' => '{{ ansistrano_release_path }}',
             ],
         ];
         $afterUpdateCodeTasks[] = [
             'name' => 'Get possible environment overrides',
             'stat' => [
-                'path' => "{{ ansistrano_release_path.stdout }}/.ddev/deploy.{{ ddev_ansible_environment }}.env.web",
+                'path' => "{{ ansistrano_release_path }}/.ddev/deploy.{{ ddev_ansible_environment }}.env.{{ item}}",
             ],
-            'register' => 'ddev_env_web_overrides',
+            'register' => 'ddev_env_overrides',
+            'loop' => '{{ ddev_services }}',
         ];
         $afterUpdateCodeTasks[] = [
-            'name' => 'Copy web env web overrides',
+            'name' => 'Enable environment overrides',
             'copy' => [
-                'src' => "{{ ansistrano_release_path.stdout }}/.ddev/deploy.{{ ddev_ansible_environment }}.env.web",
-                'dest' => "{{ ansistrano_release_path.stdout }}/.ddev/.env.web",
+                'src' => "{{ ansistrano_release_path }}/.ddev/deploy.{{ ddev_ansible_environment }}.env.{{ item.item }}",
+                'dest' => "{{ ansistrano_release_path }}/.ddev/.env.{{ item.item }}",
                 'remote_src' => true,
             ],
-            'when' => 'ddev_env_web_overrides.stat.exists',
+            'when' => 'item.stat.exists',
+            'loop' => '{{ ddev_env_overrides.results }}',
         ];
         $afterUpdateCodeTasks[] = [
             'name' => 'Get possible hostname overrides',
             'stat' => [
-                'path' => "{{ ansistrano_release_path.stdout }}/.ddev/deploy.{{ ddev_ansible_environment }}.hostname.config.yaml",
+                'path' => "{{ ansistrano_release_path }}/.ddev/deploy.{{ ddev_ansible_environment }}.hostname.config.yaml",
             ],
             'register' => 'ddev_hostname_overrides',
         ];
         $afterUpdateCodeTasks[] = [
             'name' => 'Copy hostname overrides',
             'copy' => [
-                'src' => "{{ ansistrano_release_path.stdout }}/.ddev/deploy.{{ ddev_ansible_environment }}.hostname.config.yaml",
-                'dest' => "{{ ansistrano_release_path.stdout }}/.ddev/config.hostname.yaml",
+                'src' => "{{ ansistrano_release_path }}/.ddev/deploy.{{ ddev_ansible_environment }}.hostname.config.yaml",
+                'dest' => "{{ ansistrano_release_path }}/.ddev/config.hostname.yaml",
                 'remote_src' => true,
             ],
             'when' => 'ddev_hostname_overrides.stat.exists',
@@ -158,7 +178,7 @@ class DeployGenerate
         $afterUpdateCodeTasks[] = [
             'name' => 'Register the hostname configuration',
             'ansible.builtin.slurp' => [
-                'src' => "{{ ansistrano_release_path.stdout }}/.ddev/config.hostname.yaml",
+                'src' => "{{ ansistrano_release_path }}/.ddev/config.hostname.yaml",
             ],
             'register' => 'ddev_hostname_config',
             'when' => 'ddev_hostname_overrides.stat.exists',
@@ -168,7 +188,7 @@ class DeployGenerate
             'name' => 'Configure HTTPS overrides for traefik to issue letsencrypt certificates only for additional_fqdns',
             'ansible.builtin.template' => [
                 'src' => getenv('DDEV_APPROOT') . '/.ddev/ansible/traefik-https-overrides.yaml',
-                'dest' => "{{ ansistrano_release_path.stdout }}/.ddev/traefik/config/deploy.yaml",
+                'dest' => "{{ ansistrano_release_path }}/.ddev/traefik/config/deploy.yaml",
             ],
             'when' => 'ddev_hostname_overrides.stat.exists',
         ];
@@ -176,15 +196,15 @@ class DeployGenerate
         $afterUpdateCodeTasks[] = [
             'name' => 'Get mail sending (msmtp) configuration',
             'stat' => [
-                'path' => "{{ ansistrano_release_path.stdout }}/.ddev/deploy.{{ ddev_ansible_environment }}.msmtprc",
+                'path' => "{{ ansistrano_release_path }}/.ddev/deploy.{{ ddev_ansible_environment }}.msmtprc",
             ],
             'register' => 'ddev_msmtprc',
         ];
         $afterUpdateCodeTasks[] = [
             'name' => 'Copy mail sending configuration to home',
             'copy' => [
-                'src' => "{{ ansistrano_release_path.stdout }}/.ddev/deploy.{{ ddev_ansible_environment }}.msmtprc",
-                'dest' => "{{ ansistrano_release_path.stdout }}/.ddev/homeadditions/.msmtprc",
+                'src' => "{{ ansistrano_release_path }}/.ddev/deploy.{{ ddev_ansible_environment }}.msmtprc",
+                'dest' => "{{ ansistrano_release_path }}/.ddev/homeadditions/.msmtprc",
                 'remote_src' => true,
                 'mode' => 'u=r',
             ],
@@ -193,32 +213,36 @@ class DeployGenerate
         $afterUpdateCodeTasks[] = [
             'name' => 'Enable msmtp for php',
             'copy' => [
-                'dest' => "{{ ansistrano_release_path.stdout }}/.ddev/php/msmtp.ini",
+                'dest' => "{{ ansistrano_release_path }}/.ddev/php/msmtp.ini",
                 'content' => "[PHP]\nsendmail_path = /usr/bin/msmtp -t -i"
             ],
             'when' => 'ddev_hostname_overrides.stat.exists',
         ];
 
+        if (isset($projectTypeConfig['afterUpdateCodeTasks'])) {
+            $afterUpdateCodeTasks = array_merge($afterUpdateCodeTasks, $projectTypeConfig['afterUpdateCodeTasks']);
+        }
+
         $afterUpdateCodeTasks[] = [
             'name' => 'Start the app',
             'shell' => 'ddev start',
             'args' => [
-                'chdir' => '{{ ansistrano_release_path.stdout }}',
+                'chdir' => '{{ ansistrano_release_path }}',
             ],
             'register' => 'start_result',
         ];
+
         $afterUpdateCodeTasks[] = [
             'name' => 'Log the output of starting the app',
             'copy' => [
                 'content' => "STDOUT\n{{ start_result.stdout }}\nSTDERR\n{{ start_result.stderr }}",
-                'dest' => '{{ ansistrano_release_path.stdout }}/ansible-ddev-start.log.txt',
+                'dest' => '{{ ansistrano_release_path }}/ansible-ddev-start.log.txt',
             ],
         ];
 
-        // TODO: do not hardcode drush.
         $ansibleCronConfig = [
             'name' => 'DDEV cron job for {{ ddev_project_name }}-{{ ddev_ansible_environment }}',
-            'job' => 'cd {{ ansistrano_deploy_to }}/current && ddev drush cron ',
+            'job' => 'cd {{ ansistrano_deploy_to }}/current && ' . $projectTypeConfig['cron'],
             'minute' => '{{ ddev_cron_job_minute }}',
             'hour' => '{{ ddev_cron_job_hour }}',
             'day' => '{{ ddev_cron_job_day }}',
@@ -238,16 +262,20 @@ class DeployGenerate
             $output->writeln('/.ddev/config.basin-deploy.yaml generated. Edit it to complete the server details');
         }
 
-        $ddevConfig = Yaml::parseFile(getenv('DDEV_APPROOT') . '/.ddev/config.yaml');
-        $deployEnvironmentPath = getenv('DDEV_APPROOT') . '/.ddev/deploy.' . $environment . '.env.web';
-        if (!file_exists($deployEnvironmentPath) && str_starts_with($ddevConfig['type'], 'drupal')) {
-            copy(__DIR__ . '/../Templates/deploy.drupal.env.web', $deployEnvironmentPath);
-            $output->writeln('/.ddev/deploy.' . $environment . '.env.web generated for "' . $ddevConfig['type'] . '". Edit it to complete the environment details');
+        if (isset($projectTypeConfig['envFiles'])) {
+            foreach($projectTypeConfig['envFiles'] as $suffix => $contents) {
+                $deployEnvironmentPath = getenv('DDEV_APPROOT') . '/.ddev/deploy.' . $environment . '.env.' . $suffix;
+                if (file_exists($deployEnvironmentPath)) {
+                    continue;
+                }
+                file_put_contents($deployEnvironmentPath, $contents);
+                $output->writeln('/.ddev/deploy.' . $environment . '.env.' . $suffix . ' generated for "' . $projectType . '". Edit it to complete the environment details');
+            }
         }
 
         $deployHostnamePath = getenv('DDEV_APPROOT') . '/.ddev/deploy.' . $environment . '.hostname.config.yaml';
         if (!file_exists($deployHostnamePath)) {
-            copy(__DIR__  . '/../Templates/deploy.hostname.config.yaml', $deployEnvironmentPath);
+            copy(__DIR__  . '/../Templates/deploy.hostname.config.yaml', $deployHostnamePath);
             $output->writeln('/.ddev/deploy.' . $environment . '.hostname.config.yaml generated. Edit it to complete the hostname details');
         }
         $msmtprcPath = getenv('DDEV_APPROOT') . '/.ddev/deploy.' . $environment . '.msmtprc';
@@ -257,36 +285,19 @@ class DeployGenerate
         }
 
 
-        $beforeSymlinkTasksPath = getenv('DDEV_APPROOT') . '/.ddev/deploy.before-symlink-tasks.yml';
-        // TODO: Separate Drupal specifics to another file.
-        if (str_starts_with($ddevConfig['type'], 'drupal')) {
-            $beforeSymlinkTasks[] = [
-                'name' => 'Composer install',
-                'shell' => 'ddev composer install --no-interaction --no-dev --optimize-autoloader',
-                'args' => [
-                    'chdir' => '{{ ansistrano_release_path.stdout }}',
-                ],
-            ];
-            $beforeSymlinkTasks[] = [
-                'name' => 'Deploy drupal',
-                'shell' => 'ddev drush deploy',
-                'args' => [
-                    'chdir' => '{{ ansistrano_release_path.stdout }}',
-                ],
-            ];
-
+        if (isset($projectTypeConfig['beforeSymlinkTasks'])) {
+            $beforeSymlinkTasksPath = getenv('DDEV_APPROOT') . '/.ddev/deploy.before-symlink-tasks.yml';
             $ansiblePlaybook['vars']['ansistrano_before_symlink_tasks_file'] = $beforeSymlinkTasksPath;
-
             file_put_contents(
                 $beforeSymlinkTasksPath,
                 Yaml::dump(
-                    input: $beforeSymlinkTasks ,
+                    input: $projectTypeConfig['beforeSymlinkTasks'],
                     inline: 4,
                     flags: Yaml::DUMP_MULTI_LINE_LITERAL_BLOCK
                 )
             );
 
-            $output->writeln('/.ddev/deploy.before-symlink-tasks.yaml generated for "' . $ddevConfig['type'] . '".');
+            $output->writeln('/.ddev/deploy.before-symlink-tasks.yaml generated for "' . $projectType . '".');
         }
 
         file_put_contents(
@@ -311,5 +322,24 @@ class DeployGenerate
         );
         $output->writeln('/.ddev/deploy.yaml generated');
         return Command::SUCCESS;
+    }
+
+    protected function getProjectType($ddevDockerCompose): string {
+        // Drupal, supported by DDEV
+        if (file_exists(__DIR__ . '/../ProjectType/' .  ucfirst(getenv('DDEV_PROJECT_TYPE')) . '.php')) {
+            return ucfirst(getenv('DDEV_PROJECT_TYPE'));
+        }
+
+        // Nextcloud. Capture the major version from the image declaration
+        if (
+            array_key_exists('nextcloud', $ddevDockerCompose['services']) &&
+            preg_match('/^[^:]*:(\d+)\./', $ddevDockerCompose['services']['nextcloud']['image'], $capture)
+        ) {
+            $major = $capture[1];
+            if (file_exists(__DIR__ . '/../ProjectType/Nextcloud' . $major . '.php')) {
+                return 'Nextcloud' . $major;
+            }
+        }
+        return '';
     }
 }
